requirements.txt
pytest
selenium==4.24.0
chromedriver-autoinstaller==0.6.4
geckodriver-autoinstaller==0.1.0
cryptography==41.0.4
PyYAML==6.0
retry==0.9.2
pyperclip3==0.4.1
requests==2.32.3
cryptography==41.0.4